Add template name guesser service (with suggested changes)
