+++ Resources/config/services.xml
@@ -1,0 +1,17 @@
+<?xml version="1.0" ?>
+
+<container xmlns="http://symfony.com/schema/dic/services"
+    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+    xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd">
+
+    <parameters>
+        <parameter key="sensio_framework_extra.view.guesser.class">Sensio\Bundle\FrameworkExtraBundle\Templating\TemplateGuesser</parameter>
+    </parameters>
+
+    <services>
+        <service id="sensio_framework_extra.view.guesser" class="%sensio_framework_extra.view.guesser.class%">
+            <argument type="service" id="kernel" />
+        </service>
+    </services>
+
+</container>
